DAL kapselt nun die Datenbankdaten.

+++ Server/DAL/Company.cs
@@ -8,10 +8,10 @@ namespace Server.DAL
 {
 	public class Company
 	{
-		public long ID { get; private set; }
+		public int ID { get; private set; }
 		public string Name { get; private set; }
 
-		public Company(long id, string name)
+		public Company(int id, string name)
 		{
 			ID = id;
 			Name = name;
